+++ src/UI/Cells/ExtraTypeCell.js
@@ -7,8 +7,13 @@ module.exports = NzbDroneCell.extend({
         this.$el.empty();
 
         var title = this.model.get('type');
-            this.$el.html(title);
+            this.$el.html(this.toTitleCase(title));
 
         return this;
+    },
+
+    toTitleCase : function(str)
+    {
+        return str.replace(/\w\S*/g, function(txt){return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();});
     }
 });
