+++ src/UI/Movies/Details/MoviesDetailsTemplate.hbs
@@ -43,15 +43,14 @@
         <ul class="nav nav-tabs" id="myTab">
             <li><a href="#movie-history" class="x-movie-history">History</a></li>
             <li><a href="#movie-search" class="x-movie-search">Search</a></li>
-            <li><a href="#movie-files" class="x-movie-files">Files</a></li>
-            <li><a href="#movie-extra-files" class="x-movie-extra-files">Extras</a></li>
+            <li><a href="#movie-files-tabs" class="x-movie-files-tabs">Files</a></li>
             <li><a href="#movie-titles" class="x-movie-titles">Titles</a></li>
         </ul>
+
         <div class="tab-content">
             <div class="tab-pane" id="movie-history"/>
             <div class="tab-pane" id="movie-search"/>
-            <div class="tab-pane" id="movie-files"/>
-            <div class="tab-pane" id="movie-extra-files"/>
+            <div class="tab-pane" id="movie-files-tabs"/>
             <div class="tab-pane" id="movie-titles"/>
         </div>
       </div>
