--- src/UI/Movies/Extras/ExtraFilesLayoutTemplate.hbs
@@ -1,3 +1,0 @@
-<div id="movie-extra-files-region">
-    <div id="movie-extra-files-grid" class="table-responsive"></div>
-</div>
