+++ src/UI/Movies/Files/AllFilesLayoutTemplate.hbs
@@ -1,0 +1,5 @@
+<div class="x-movie-files" id="movie-files">
+    <div id="movie-media-files" />
+    <legend>Extras</legend>
+    <div id="movie-extra-files" />
+</div>
