+++ src/UI/Movies/Files/Extras/ExtraFilesLayoutTemplate.hbs
@@ -1,0 +1,1 @@
+<div class="extra-files-table table-responsive"></div>
