--- src/UI/Movies/Files/FilesLayoutTemplate.hbs
@@ -1,3 +1,0 @@
-<div id="movie-files-region">
-    <div id="movie-files-grid" class="table-responsive"></div>
-</div>
