+++ src/UI/Movies/Files/Media/FilesLayoutTemplate.hbs
@@ -1,0 +1,2 @@
+<div id="movie-files-grid" class="table-responsive"/>
+
