+++ src/UI/Shared/Modal/ModalController.js
@@ -11,7 +11,7 @@ var RenamePreviewLayout = require('../../Rename/RenamePreviewLayout');
 var ManualImportLayout = require('../../ManualImport/ManualImportLayout');
 var FileBrowserLayout = require('../FileBrowser/FileBrowserLayout');
 var MoviesDetailsLayout = require('../../Movies/Details/MoviesDetailsLayout');
-var EditFileView = require("../../Movies/Files/Edit/EditFileView");
+var EditFileView = require("../../Movies/Files/Media/Edit/EditFileView");
 
 module.exports = Marionette.AppRouter.extend({
 		initialize : function() {
